fileConverter/requirements.txt
pillow
pillow-heif
opencv-python
numpy

